+++ neptune/new/internal/backends/utils.py
@@ -19,24 +19,25 @@ import os
 import socket
 import sys
 import time
+from functools import lru_cache, wraps
 from typing import Optional, Dict, TYPE_CHECKING
-
 from urllib.parse import urlparse, urljoin
+
 
 import click
 import requests
-
+import urllib3
+from urllib3.exceptions import NewConnectionError
 from bravado.client import SwaggerClient
 from bravado.exception import BravadoConnectionError, BravadoTimeoutError, HTTPForbidden, \
     HTTPServerError, HTTPUnauthorized, HTTPServiceUnavailable, HTTPRequestTimeout, \
-    HTTPGatewayTimeout, HTTPBadGateway, HTTPClientError, HTTPTooManyRequests
+    HTTPGatewayTimeout, HTTPBadGateway, HTTPClientError, HTTPTooManyRequests, HTTPError
 from bravado.http_client import HttpClient
 from bravado_core.formatter import SwaggerFormat
 from packaging.version import Version
 from requests import Session
-from urllib3.exceptions import NewConnectionError
 
-from neptune.new.envs import NEPTUNE_RETRIES_TIMEOUT_ENV
+from neptune.new.envs import NEPTUNE_RETRIES_TIMEOUT_ENV, NEPTUNE_ALLOW_SELF_SIGNED_CERTIFICATE
 from neptune.new.exceptions import SSLError, NeptuneConnectionLostException, \
     Unauthorized, Forbidden, CannotResolveHostname, UnsupportedClientVersion, ClientHttpError, \
     NeptuneFeaturesNotAvailableException
@@ -110,6 +111,7 @@ def with_api_exceptions_handler(func):
     return wrapper
 
 
+@lru_cache(maxsize=None, typed=True)
 def verify_host_resolution(url: str) -> None:
     host = urlparse(url).netloc.split(':')[0]
     try:
@@ -177,3 +179,40 @@ class MissingApiClient(SwaggerClient):
 
     def __getattr__(self, item):
         raise NeptuneFeaturesNotAvailableException(features=self._backend.missing_features)
+
+
+# https://stackoverflow.com/a/44776960
+def cache(func):
+    """
+    Transform mutable dictionary into immutable before call to lru_cache
+    """
+    class HDict(dict):
+        def __hash__(self):
+            return hash(frozenset(self.items()))
+
+    func = lru_cache(maxsize=None, typed=True)(func)
+
+    @wraps(func)
+    def wrapper(*args, **kwargs):
+        args = tuple([HDict(arg) if isinstance(arg, dict) else arg for arg in args])
+        kwargs = {k: HDict(v) if isinstance(v, dict) else v for k, v in kwargs.items()}
+        return func(*args, **kwargs)
+
+    wrapper.cache_clear = func.cache_clear
+    return wrapper
+
+
+def ssl_verify():
+    if os.getenv(NEPTUNE_ALLOW_SELF_SIGNED_CERTIFICATE):
+        urllib3.disable_warnings()
+        return False
+
+    return True
+
+
+def parse_validation_errors(error: HTTPError) -> Dict[str, str]:
+    return {
+        f"{error_description.get('errorCode').get('name')}": error_description.get('context', '')
+        for validation_error in error.swagger_result.validationErrors
+        for error_description in validation_error.get('errors')
+    }